requirements.txt
aiohttp==3.6.2
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.1.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-storages==1.10.1
djangorestframework==3.11.0
google-api-core==1.22.2
google-auth==1.22.0
google-cloud-core==1.4.1
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
mysqlclient==1.4.6
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pylint==2.5.3
python-resize-image==1.1.19
pytz==2020.1
requests==2.24.0
rsa==4.6
sentry-sdk==0.17.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
yarl==1.6.0
